// src/client/singlePlayer/types/singlePlayerStep.ts
import SinglePlayerAction from "./singlePlayerAction";
import SinglePlayerStepTransitionRule from "./singlePlayerStepTransitionRule";

export default interface SinglePlayerStep
{
    startDelay: number;
    actionsOnStart: SinglePlayerAction[];
    transitionRules: SinglePlayerStepTransitionRule[];
    actionsOnEnd: SinglePlayerAction[];
}